Combined the bot_menu and start_conversation files. Replaced the "start" command with a constant that was previously created, but was not used in the project.
