+++ src/jobs.py
@@ -1,0 +1,22 @@
+from telegram.ext import CallbackContext
+
+
+async def weekly_stat_job(context: CallbackContext) -> None:
+    """
+    Send weekly statistics on the number of requests in the work
+    """
+
+
+async def monthly_receipt_reminder_job(context: CallbackContext) -> None:
+    """
+    Send monthly reminder about the receipt formation during payment
+    Only for self-employed users
+    """
+
+
+async def monthly_stat_job(context: CallbackContext) -> None:
+    """
+    Send monthly statistics on the number of successfully
+    closed requests.
+    Only if the user had requests
+    """
